Move DapProblemType to sub-module
